rudimentary sync implementation & fixed compile issues

+++ oneEngine/oneGame/source/renderer/gpuw/Textures.cpp
@@ -113,7 +113,7 @@ int gpu::Texture::free ( void )
 
 int gpu::Texture::upload ( gpu::Buffer& buffer, const uint level )
 {
-	glBindBuffer(GL_PIXEL_UNPACK_BUFFER_ARB, buffer.nativePtr());
+	glBindBuffer(GL_PIXEL_UNPACK_BUFFER_ARB, (GLuint)buffer.nativePtr());
 	glTextureSubImage2D(m_texture, level, 0, 0, m_width, m_height, m_glformat, m_gltype, NULL);
 	glBindBuffer(GL_PIXEL_UNPACK_BUFFER_ARB, 0);
